refactor(pdf): isolate visual source geometry
